backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
python-multipart
pydantic
opencv-python-headless
ultralytics
numpy
